cuoi tuan

+++ src/main/java/qms/bvp/common/ReceptionStatus.java
@@ -4,8 +4,8 @@ package qms.bvp.common;
  * Created by Admin on 8/17/2018.
  */
 public class ReceptionStatus {
-    public static final int DangChoTiepDon=0;
-    public static final int DangTiepDon=1;
-    public static final int DaTiepDon=2;
-    public static final int BoQua=3;
+    public static final byte DangChoTiepDon=0;
+    public static final byte DangTiepDon=1;
+    public static final byte DaTiepDon=2;
+    public static final byte BoQua=3;
 }
